chore(arch): enforce boundary rules, relocate usePopupQueue, add ADR-005, document mobile workspace
